setup.sh was updated with condition

+++ setup.sh
@@ -20,8 +20,16 @@ sudo apt-get autoremove -y
 sudo curl -o /usr/local/bin/docker-compose -L "https://github.com/docker/compose/releases/download/1.8.1/docker-compose-$(uname -s)-$(uname -m)"
 sudo chmod +x /usr/local/bin/docker-compose
 sudo apt-get install -y python3-pip
-export LC_ALL=C >> ~/.bashrc
+export LC_ALL=C
 source ~/.bashrc
 sudo pip3 install --upgrade pip
 sudo -H pip3 install fabric3 django-environ
-sudo reboot
+echo "You need to reboot the system"
+echo "Type 'yes' to continue, or 'no' to abort:"
+read res
+if ( "$res" == "yes" ); then
+    ehco "System is going for restart"
+    sudo reboot
+else
+    echo "Aborted, but remember to restart your system to use the upgraded packaged"
+fi
